[#5] Refactor : model폴더 추가 후 구조 수정
