+++ src/main/java/org/goormthon/seasonthon/nocheongmaru/domain/feed/repository/feedLike/FeedLikeJpaRepository.java
@@ -1,6 +1,6 @@
 package org.goormthon.seasonthon.nocheongmaru.domain.feed.repository.feedLike;
 
-import org.goormthon.seasonthon.nocheongmaru.domain.feed.entity.FeedLike;
+import org.goormthon.seasonthon.nocheongmaru.domain.feed.model.entity.FeedLike;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FeedLikeJpaRepository extends JpaRepository<FeedLike, Long> {
